+++ R/helpers.R
@@ -46,18 +46,6 @@ is.historical <- function(.data) { # nolint: object_name_linter
 }
 
 
-
-#' not-in operator
-#' @inheritParams base::`%in%`
-#' @examples
-#' 2 %notin% c(1,3) # TRUE
-#' @return A logical vector, indicating if a match was NOT located for each element of x
-#' @export
-`%notin%` <- function(x, table) {
-  return(!(x %in% table))
-}
-
-
 #' checkmate helper: Assert "generic" data.table/data.frame/tbl/tibble type
 #' @param .data Object to test if is data.table, data.frame, tbl or tibble
 #' @param ...   Parameters passed to checkmate::check_*
@@ -65,12 +53,13 @@ is.historical <- function(.data) { # nolint: object_name_linter
 #' @inherit checkmate::assert return
 #' @noRd
 assert_data_like <- function(.data, ..., add = NULL) {
-  checkmate::assert( # nolint start: indentation_linter
+  checkmate::assert(
     checkmate::check_class(.data, "tbl_dbi", ...),
     checkmate::check_data_frame(.data, ...),
     checkmate::check_data_table(.data, ...),
     checkmate::check_tibble(.data, ...),
-    add = add) # nolint end
+    add = add
+  )
 }
 
 
@@ -81,11 +70,12 @@ assert_data_like <- function(.data, ..., add = NULL) {
 #' @inherit checkmate::assert return
 #' @noRd
 assert_timestamp_like <- function(timestamp, ..., add = NULL) {
-  checkmate::assert( # nolint start: indentation_linter
+  checkmate::assert(
     checkmate::check_posixct(timestamp, ...),
     checkmate::check_character(timestamp, ...),
     checkmate::check_date(timestamp, ...),
-    add = add) # nolint end
+    add = add
+  )
 }
 
 
@@ -96,11 +86,12 @@ assert_timestamp_like <- function(timestamp, ..., add = NULL) {
 #' @inherit checkmate::assert return
 #' @noRd
 assert_dbtable_like <- function(db_table, ..., add = NULL) {
-  checkmate::assert( # nolint start: indentation_linter
+  checkmate::assert(
     checkmate::check_character(db_table, pattern = r"{^\w*.\w*$}", ...),
     checkmate::check_class(db_table, "Id", ...),
     checkmate::check_class(db_table, "tbl_dbi", ...),
-    add = add) # nolint end
+    add = add
+  )
 }
 
 
@@ -111,9 +102,10 @@ assert_dbtable_like <- function(db_table, ..., add = NULL) {
 #' @inherit checkmate::assert return
 #' @noRd
 assert_id_like <- function(id, ..., add = NULL) {
-  checkmate::assert( # nolint start: indentation_linter
+  checkmate::assert(
     checkmate::check_character(id, ...),
     checkmate::check_class(id, "Id", ...),
     checkmate::check_class(id, "tbl_dbi", ...),
-    add = add) # nolint end
+    add = add
+  )
 }